requirements.txt
essentia>=2.1b6.dev1389,<2.2
fastapi>=0.115,<0.116
uvicorn[standard]>=0.30,<0.31
python-dotenv>=1.0,<2.0
